requirements_infer.txt
rknnlite2
numpy
pillow
transformers
torch
